requirements.txt
streamlit
pdfminer.six
PyMuPDF
pandas
Pillow
streamlit-paste-button
xlsxwriter
